--- packages/core/src/provider/index.ts
@@ -1,4 +1,0 @@
-export * from "./text";
-export * from "./utils";
-export * from "./interfaces";
-export * from "./image";
